Only evaluate PASS variants to avoid crash
hap.py has a preprocessing bug
<https://github.com/Illumina/hap.py/issues/37> that causes a segfault
that can be avoided if you don't have any 0/0 calls in the sample.
DeepVariant filters all its 0/0 calls (and some other calls), and the
official DV case studies use --pass-only in their evaluations
<https://github.com/google/deepvariant/blob/r1.6.1/docs/deepvariant-pacbio-model-case-study.md#benchmark-output>.
So this adds --pass-only to the evaluation here to work around the bug
and match the DV evals more closely.